requirements.txt
discord.py[speed]
motor==3.0.0
PyNaCl
dnspython
python-dotenv
aiohttp
youtube_dl